Refactor: Update favorite coffee loading to use streams (#1)
* Refactor: Update favorite coffee loading to use streams

This commit refactors the way favorite coffees are loaded and updated in the application.

Previously, favorite coffees were loaded once and then refreshed when changes occurred in the `FavouriteCoffeeBloc`.

Now, the `LoadFavoriteCoffeesUseCase` and `CoffeeRepository` return a `Stream` of favorite coffees. This allows the `CoffeeBloc` and `FavouriteCoffeeBloc` to subscribe to these streams and automatically update their state when the list of favorite coffees changes in the underlying storage.

Key changes:
- `LoadFavoriteCoffeesUseCase.execute` now returns `Stream<Either<CoffeeFailure, List<Coffee>>>`.
- `CoffeeRepository.getFavoriteCoffees` now returns `Stream<List<Map<String, dynamic>>>` and listens to `GetStorage` key changes to emit new lists.
- `CoffeeBloc` and `FavouriteCoffeeBloc` now use `emit.forEach` to handle the stream of favorite coffees.
- Introduced `CoffeeStatus.loadingFavourites` to differentiate between loading a new coffee image and loading/updating the favorites list.
- Removed the explicit `RefreshFavouriteCoffeeEvent` listener from `CoffeeIndividualWidget` as updates are now handled reactively by the stream.
- Updated tests to reflect the new stream-based approach.

* Deactivate semantic PR check

This commit removes the `semantic-pull-request` job from the GitHub Actions workflow. This job was previously used to enforce semantic commit messages on pull requests.

* Fix: Specify Flutter version and exclude coverage for injection

This commit updates the GitHub Actions workflow to:

- Specify `flutter_version: 3.32.8`.
- Exclude files in `**/injection/**` from code coverage calculation.

* Fix: Exclude routes from coverage

This commit updates the GitHub Actions workflow to exclude files in the `**/routes/` directory from code coverage calculation. This is likely done because these files contain generated code or boilerplate that doesn't require unit testing.

* Set minimum code coverage to 98%

This commit updates the `very_good_workflows` action in the `main.yaml` GitHub workflow to enforce a minimum code coverage of 98%.